requirements.txt
# WoodMind / diploma — веб-приложение и загрузка данных
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0
openpyxl>=3.1.0
pyvis>=0.3.2

# Опционально: method_product/run_analysis.py (docx, графики)
python-docx>=1.1.0
matplotlib>=3.7.0
